+++ src/commands/serve-web-app.js
@@ -1,32 +1,9 @@
-import path from 'path'
-
-import glob from 'glob'
-
-import {getDefaultHTMLConfig} from '../appConfig'
+import createServeWebAppConfig from '../createServeWebAppConfig'
 import webpackServer from '../webpackServer'
 
+/**
+ * Serve a plain JS app.
+ */
 export default function serveWebApp(args, cb) {
-  let entry = path.resolve(args._[1] || 'src/index.js')
-  let dist = path.resolve(args._[2] || 'dist')
-
-  let config = {
-    babel: {
-      commonJSInterop: true,
-    },
-    entry: [entry],
-    output: {
-      path: dist,
-      filename: 'app.js',
-      publicPath: '/',
-    },
-    plugins: {
-      html: getDefaultHTMLConfig(),
-    },
-  }
-
-  if (glob.sync('public/').length !== 0) {
-    config.plugins.copy = [{from: path.resolve('public'), to: dist, ignore: '.gitkeep'}]
-  }
-
-  webpackServer(args, config, cb)
+  webpackServer(args, createServeWebAppConfig(args), cb)
 }