+++ tests/commands/serve-test.js
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 
-import {spawn} from 'cross-spawn'
+import spawn from 'cross-spawn'
 import EventSource from 'eventsource'
 import expect from 'expect'
 import rimraf from 'rimraf'